(maint) Create a top level tasks validator
