+++ spec/acceptance/validate_metadata_spec.rb
@@ -126,79 +126,4 @@ describe 'Running metadata validation' do
       end
     end
   end
-
-  context 'with valid task' do
-    include_context 'in a new module', 'foo'
-
-    before(:all) do
-      File.open(File.join('tasks', 'valid.json'), 'w') do |f|
-        f.puts <<-EOS
-{
-  "puppet_task_version": 1,
-  "supports_noop": false,
-  "description": "A short description of this task"
-}
-        EOS
-      end
-    end
-
-    describe command('pdk validate metadata') do
-      its(:exit_status) { is_expected.to eq(0) }
-      its(:stderr) { is_expected.to match(metadata_syntax_spinner) }
-      its(:stderr) { is_expected.to match(module_style_spinner) }
-      its(:stderr) { is_expected.to match(task_style_spinner) }
-    end
-
-    describe command('pdk validate metadata --format junit') do
-      its(:exit_status) { is_expected.to eq(0) }
-      its(:stderr) { is_expected.to match(metadata_syntax_spinner) }
-      its(:stderr) { is_expected.to match(module_style_spinner) }
-      its(:stderr) { is_expected.to match(task_style_spinner) }
-      its(:stdout) { is_expected.to pass_validation(junit_xsd) }
-
-      its(:stdout) { is_expected.to have_junit_testsuite('task-metadata-lint') }
-    end
-  end
-
-  context 'with invalid task' do
-    include_context 'in a new module', 'foo'
-
-    before(:all) do
-      File.open(File.join('tasks', 'invalid.json'), 'w') do |f|
-        f.puts <<-EOS
-{
-  "puppet_task_version": 1,
-  "supports_noop": "false",
-  "description": "A short description of this task"
-}
-        EOS
-      end
-    end
-
-    describe command('pdk validate metadata') do
-      its(:exit_status) { is_expected.not_to eq(0) }
-      its(:stderr) { is_expected.to match(metadata_syntax_spinner) }
-      its(:stderr) { is_expected.to match(module_style_spinner) }
-      its(:stderr) { is_expected.to match(task_style_spinner) }
-      its(:stdout) { is_expected.to match(%r{The property '#/supports_noop' of type string did not match the following type: boolean}i) }
-    end
-
-    describe command('pdk validate metadata --format junit') do
-      its(:exit_status) { is_expected.not_to eq(0) }
-      its(:stderr) { is_expected.to match(metadata_syntax_spinner) }
-      its(:stderr) { is_expected.to match(module_style_spinner) }
-      its(:stderr) { is_expected.to match(task_style_spinner) }
-      its(:stdout) { is_expected.to pass_validation(junit_xsd) }
-
-      its(:stdout) do
-        is_expected.to have_junit_testcase.in_testsuite('task-metadata-lint').with_attributes(
-          'classname' => 'task-metadata-lint',
-          'name'      => a_string_matching(%r{invalid.json}),
-        ).that_failed(
-          'type'    => a_string_matching(%r{error}i),
-          'message' => a_string_matching(%r{The property '#/supports_noop' of type string did not match the following type: boolean}i),
-        )
-      end
-    end
-  end
 end
