+++ spec/acceptance/validate_tasks_spec.rb
@@ -1,0 +1,75 @@
+require 'spec_helper_acceptance'
+
+describe 'Running tasks validation' do
+  let(:junit_xsd) { File.join(RSpec.configuration.fixtures_path, 'JUnit.xsd') }
+  let(:task_style_spinner) { %r{checking task metadata style}i }
+
+  include_context 'with a fake TTY'
+
+  context 'with valid task' do
+    include_context 'in a new module', 'foo'
+
+    before(:all) do
+      File.open(File.join('tasks', 'valid.json'), 'w') do |f|
+        f.puts <<-EOS
+{
+  "puppet_task_version": 1,
+  "supports_noop": false,
+  "description": "A short description of this task"
+}
+        EOS
+      end
+    end
+
+    describe command('pdk validate tasks') do
+      its(:exit_status) { is_expected.to eq(0) }
+      its(:stderr) { is_expected.to match(task_style_spinner) }
+    end
+
+    describe command('pdk validate tasks --format junit') do
+      its(:exit_status) { is_expected.to eq(0) }
+      its(:stderr) { is_expected.to match(task_style_spinner) }
+      its(:stdout) { is_expected.to pass_validation(junit_xsd) }
+
+      its(:stdout) { is_expected.to have_junit_testsuite('task-metadata-lint') }
+    end
+  end
+
+  context 'with invalid task' do
+    include_context 'in a new module', 'foo'
+
+    before(:all) do
+      File.open(File.join('tasks', 'invalid.json'), 'w') do |f|
+        f.puts <<-EOS
+{
+  "puppet_task_version": 1,
+  "supports_noop": "false",
+  "description": "A short description of this task"
+}
+        EOS
+      end
+    end
+
+    describe command('pdk validate tasks') do
+      its(:exit_status) { is_expected.not_to eq(0) }
+      its(:stderr) { is_expected.to match(task_style_spinner) }
+      its(:stdout) { is_expected.to match(%r{The property '#/supports_noop' of type string did not match the following type: boolean}i) }
+    end
+
+    describe command('pdk validate tasks --format junit') do
+      its(:exit_status) { is_expected.not_to eq(0) }
+      its(:stderr) { is_expected.to match(task_style_spinner) }
+      its(:stdout) { is_expected.to pass_validation(junit_xsd) }
+
+      its(:stdout) do
+        is_expected.to have_junit_testcase.in_testsuite('task-metadata-lint').with_attributes(
+          'classname' => 'task-metadata-lint',
+          'name'      => a_string_matching(%r{invalid.json}),
+        ).that_failed(
+          'type'    => a_string_matching(%r{error}i),
+          'message' => a_string_matching(%r{The property '#/supports_noop' of type string did not match the following type: boolean}i),
+        )
+      end
+    end
+  end
+end
