+++ spec/spec_helper.rb
@@ -65,6 +65,7 @@ RSpec.shared_context :validators do
       PDK::Validate::MetadataValidator,
       PDK::Validate::PuppetValidator,
       PDK::Validate::RubyValidator,
+      PDK::Validate::TasksValidator,
     ]
   end
 end
